e2e: combine OpenStack advanced tests within one Nodepool
This will make our tests to run faster, especially if we keep adding
tests.
