+++ test/e2e/nodepool_osp_advanced_test.go
@@ -26,9 +26,14 @@ const (
 	// For now we hardcode the network name, but we should make it configurable
 	// and maybe use Gophercloud to create a network with a dynamic name.
 	additionalNetworkName = "hcp-nodepool-multinet-e2e"
+
+	// The default availability zone for OpenStack is "nova" but the AZ name
+	// can be different depending on the OpenStack deployment so it can be
+	// overridden by the e2e user.
+	defaultAvailabilityZone = "nova"
 )
 
-type OpenStackMultinetTest struct {
+type OpenStackAdvancedTest struct {
 	DummyInfraSetup
 	ctx                         context.Context
 	managementClient            crclient.Client
@@ -36,8 +41,8 @@ type OpenStackMultinetTest struct {
 	hostedControlPlaneNamespace string
 }
 
-func NewOpenStackMultinetTest(ctx context.Context, mgmtClient crclient.Client, hostedCluster *hyperv1.HostedCluster) *OpenStackMultinetTest {
-	return &OpenStackMultinetTest{
+func NewOpenStackAdvancedTest(ctx context.Context, mgmtClient crclient.Client, hostedCluster *hyperv1.HostedCluster) *OpenStackAdvancedTest {
+	return &OpenStackAdvancedTest{
 		ctx:                         ctx,
 		hostedCluster:               hostedCluster,
 		managementClient:            mgmtClient,
@@ -45,20 +50,20 @@ func NewOpenStackMultinetTest(ctx context.Context, mgmtClient crclient.Client, h
 	}
 }
 
-func (o OpenStackMultinetTest) Setup(t *testing.T) {
-	t.Log("Starting test OpenStackMultinetTest")
+func (o OpenStackAdvancedTest) Setup(t *testing.T) {
+	t.Log("Starting test OpenStackAdvancedTest")
 
 	if globalOpts.Platform != hyperv1.OpenStackPlatform {
 		t.Skip("test only supported on platform OpenStack")
 	}
 
-	// The feature that is being tested here is only available in 4.18+
+	// The features that are being tested here is only available in 4.18+
 	if e2eutil.IsLessThan(e2eutil.Version418) {
 		t.Skip("test only applicable for 4.18+")
 	}
 }
 
-func (o OpenStackMultinetTest) Run(t *testing.T, nodePool hyperv1.NodePool, _ []corev1.Node) {
+func (o OpenStackAdvancedTest) Run(t *testing.T, nodePool hyperv1.NodePool, _ []corev1.Node) {
 	np := &hyperv1.NodePool{}
 	e2eutil.EventuallyObject(t, o.ctx, "NodePool to have additional networks configured",
 		func(ctx context.Context) (*hyperv1.NodePool, error) {
@@ -81,6 +86,22 @@ func (o OpenStackMultinetTest) Run(t *testing.T, nodePool hyperv1.NodePool, _ []
 					},
 				}, ptr.Deref(np.Spec.Platform.OpenStack, hyperv1.OpenStackNodePoolPlatform{}).AdditionalPorts)
 				return diff == "", fmt.Sprintf("incorrect additional networks: %v", diff), nil
+			},
+		},
+	)
+	e2eutil.EventuallyObject(t, o.ctx, "NodePool to have availability zone configured",
+		func(ctx context.Context) (*hyperv1.NodePool, error) {
+			err := o.managementClient.Get(ctx, util.ObjectKey(&nodePool), np)
+			return np, err
+		},
+		[]e2eutil.Predicate[*hyperv1.NodePool]{
+			func(nodePool *hyperv1.NodePool) (done bool, reasons string, err error) {
+				want, got := hyperv1.OpenStackPlatform, nodePool.Spec.Platform.Type
+				return want == got, fmt.Sprintf("expected NodePool to have platform %s, got %s", want, got), nil
+			},
+			func(pool *hyperv1.NodePool) (done bool, reasons string, err error) {
+				diff := cmp.Diff(getAZName(), ptr.Deref(np.Spec.Platform.OpenStack, hyperv1.OpenStackNodePoolPlatform{}).AvailabilityZone)
+				return diff == "", fmt.Sprintf("incorrect availability zone: %v", diff), nil
 			},
 		},
 	)
@@ -112,18 +133,44 @@ func (o OpenStackMultinetTest) Run(t *testing.T, nodePool hyperv1.NodePool, _ []
 			},
 		},
 	)
+	e2eutil.EventuallyObjects(t, o.ctx, "CAPI Machines to be created with the correct availability zone",
+		func(ctx context.Context) ([]*capiv1.Machine, error) {
+			list := &capiv1.MachineList{}
+			err := o.managementClient.List(ctx, list, crclient.InNamespace(o.hostedControlPlaneNamespace), crclient.MatchingLabels{capiv1.MachineDeploymentNameLabel: nodePool.Name})
+			oms := make([]*capiv1.Machine, len(list.Items))
+			for i := range list.Items {
+				oms[i] = &list.Items[i]
+			}
+			return oms, err
+		},
+		[]e2eutil.Predicate[[]*capiv1.Machine]{
+			func(machines []*capiv1.Machine) (done bool, reasons string, err error) {
+				return len(machines) == int(*nodePool.Spec.Replicas), fmt.Sprintf("expected %d Machines, got %d", *nodePool.Spec.Replicas, len(machines)), nil
+			},
+		},
+		[]e2eutil.Predicate[*capiv1.Machine]{
+			func(machine *capiv1.Machine) (done bool, reasons string, err error) {
+				if machine.Spec.FailureDomain == nil {
+					return false, "Machine does not have a failure domain", nil
+				}
+				want, got := getAZName(), *machine.Spec.FailureDomain
+				return want == got, fmt.Sprintf("expected Machine to have failure domain %s, got %s", want, got), nil
+			},
+		},
+	)
 }
 
-func (o OpenStackMultinetTest) BuildNodePoolManifest(defaultNodepool hyperv1.NodePool) (*hyperv1.NodePool, error) {
+func (o OpenStackAdvancedTest) BuildNodePoolManifest(defaultNodepool hyperv1.NodePool) (*hyperv1.NodePool, error) {
 	nodePool := &hyperv1.NodePool{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      o.hostedCluster.Name + "-" + "test-osp-multinet",
+			Name:      o.hostedCluster.Name + "-" + "test-osp-advanced",
 			Namespace: o.hostedCluster.Namespace,
 		},
 	}
 	defaultNodepool.Spec.DeepCopyInto(&nodePool.Spec)
 
 	nodePool.Spec.Replicas = &oneReplicas
+	nodePool.Spec.Platform.OpenStack.AvailabilityZone = getAZName()
 	nodePool.Spec.Platform.OpenStack.AdditionalPorts = []hyperv1.PortSpec{
 		{
 			Network: &hyperv1.NetworkParam{
@@ -137,10 +184,17 @@ func (o OpenStackMultinetTest) BuildNodePoolManifest(defaultNodepool hyperv1.Nod
 	return nodePool, nil
 }
 
-func (o OpenStackMultinetTest) SetupInfra(t *testing.T) error {
+func getAZName() string {
+	if globalOpts.ConfigurableClusterOptions.OpenStackNodeAvailabilityZone != "" {
+		return globalOpts.ConfigurableClusterOptions.OpenStackNodeAvailabilityZone
+	}
+	return defaultAvailabilityZone
+}
+
+func (o OpenStackAdvancedTest) SetupInfra(t *testing.T) error {
 	return nil
 }
 
-func (o OpenStackMultinetTest) TeardownInfra(t *testing.T) error {
+func (o OpenStackAdvancedTest) TeardownInfra(t *testing.T) error {
 	return nil
 }
